# Faster integer hashing (fixes #37800 UB)