@jmkuhn 2020-10-14
With this PR you have
```
julia> hash(0.0) == hash(-0.0)
true
```