@vtjnash 2020-10-15
Ah, you're right. Good catch. I completely botched that and forgot to use `isequal` with hashing instead of `==`.